+++ src/main/java/com/pragma/plazoleta/PlazoletaApplication.java
@@ -1,0 +1,13 @@
+package com.pragma.plazoleta;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PlazoletaApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PlazoletaApplication.class, args);
+	}
+
+}
