+++ src/main/java/com/pragma/plazoleta/application/dto/request/ObjectRequestDto.java
@@ -1,0 +1,10 @@
+package com.pragma.plazoleta.application.dto.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ObjectRequestDto {
+    private String name;
+}
