+++ src/main/java/com/pragma/plazoleta/application/dto/response/ObjectResponseDto.java
@@ -1,0 +1,11 @@
+package com.pragma.plazoleta.application.dto.response;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ObjectResponseDto {
+    private Long id;
+    private String name;
+}
