+++ src/main/java/com/pragma/plazoleta/application/mapper/IObjectRequestMapper.java
@@ -1,0 +1,13 @@
+package com.pragma.plazoleta.application.mapper;
+
+import com.pragma.plazoleta.application.dto.request.ObjectRequestDto;
+import com.pragma.plazoleta.domain.model.ObjectModel;
+import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
+
+@Mapper(componentModel = "spring",
+        unmappedTargetPolicy = ReportingPolicy.IGNORE,
+        unmappedSourcePolicy = ReportingPolicy.IGNORE)
+public interface IObjectRequestMapper {
+    ObjectModel toObject(ObjectRequestDto objectRequestDto);
+}
