+++ src/main/java/com/pragma/plazoleta/application/mapper/IObjectResponseMapper.java
@@ -1,0 +1,17 @@
+package com.pragma.plazoleta.application.mapper;
+
+import com.pragma.plazoleta.application.dto.response.ObjectResponseDto;
+import com.pragma.plazoleta.domain.model.ObjectModel;
+import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
+
+import java.util.List;
+
+@Mapper(componentModel = "spring",
+        unmappedTargetPolicy = ReportingPolicy.IGNORE,
+        unmappedSourcePolicy = ReportingPolicy.IGNORE)
+public interface IObjectResponseMapper {
+    ObjectResponseDto toResponse(ObjectModel objectModel);
+
+    List<ObjectResponseDto> toResponseList(List<ObjectModel> objectModelList);
+}
