+++ src/main/java/com/pragma/plazoleta/domain/api/IObjectServicePort.java
@@ -1,0 +1,12 @@
+package com.pragma.plazoleta.domain.api;
+
+import com.pragma.plazoleta.domain.model.ObjectModel;
+
+import java.util.List;
+
+public interface IObjectServicePort {
+
+    void saveObject(ObjectModel objectModel);
+
+    List<ObjectModel> getAllObjects();
+}
