+++ src/main/java/com/pragma/plazoleta/domain/exception/DomainException.java
@@ -1,0 +1,7 @@
+package com.pragma.plazoleta.domain.exception;
+
+public class DomainException extends RuntimeException {
+    public DomainException(String message) {
+        super(message);
+    }
+}
