+++ src/main/java/com/pragma/plazoleta/domain/model/ObjectModel.java
@@ -1,0 +1,15 @@
+package com.pragma.plazoleta.domain.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class ObjectModel {
+    private Long id;
+    private String name;
+}
