+++ src/main/java/com/pragma/plazoleta/domain/spi/IObjectPersistencePort.java
@@ -1,0 +1,10 @@
+package com.pragma.plazoleta.domain.spi;
+
+import com.pragma.plazoleta.domain.model.ObjectModel;
+import java.util.List;
+
+public interface IObjectPersistencePort {
+    ObjectModel saveObject(ObjectModel objectModel);
+
+    List<ObjectModel> getAllObjects();
+}
