+++ src/main/java/com/pragma/plazoleta/domain/usecase/ObjectUseCase.java
@@ -1,0 +1,26 @@
+package com.pragma.plazoleta.domain.usecase;
+
+import com.pragma.plazoleta.domain.api.IObjectServicePort;
+import com.pragma.plazoleta.domain.model.ObjectModel;
+import com.pragma.plazoleta.domain.spi.IObjectPersistencePort;
+
+import java.util.List;
+
+public class ObjectUseCase implements IObjectServicePort {
+
+    private final IObjectPersistencePort objectPersistencePort;
+
+    public ObjectUseCase(IObjectPersistencePort objectPersistencePort) {
+        this.objectPersistencePort = objectPersistencePort;
+    }
+
+    @Override
+    public void saveObject(ObjectModel objectModel) {
+        objectPersistencePort.saveObject(objectModel);
+    }
+
+    @Override
+    public List<ObjectModel> getAllObjects() {
+        return objectPersistencePort.getAllObjects();
+    }
+}
