+++ src/main/java/com/pragma/plazoleta/infrastructure/documentation/OpenApiConfiguration.java
@@ -1,0 +1,27 @@
+package com.pragma.plazoleta.infrastructure.documentation;
+
+import io.swagger.v3.oas.models.Components;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.info.License;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class OpenApiConfiguration {
+
+    @Bean
+    public OpenAPI customOpenApi(@Value("${appdescription}") String appDescription,
+                                 @Value("${appversion}") String appVersion){
+        return new OpenAPI()
+            .components(new Components())
+            .info(new Info()
+                .title("Hexagonal Power-up API")
+                .version(appVersion)
+                .description(appDescription)
+                .termsOfService("http://swagger.io/terms/")
+                .license(new License().name("Apache 2.0").url("http://springdoc.org"))
+            );
+    }
+}
