+++ src/main/java/com/pragma/plazoleta/infrastructure/exception/NoDataFoundException.java
@@ -1,0 +1,7 @@
+package com.pragma.plazoleta.infrastructure.exception;
+
+public class NoDataFoundException extends RuntimeException {
+    public NoDataFoundException() {
+        super();
+    }
+}
