+++ src/main/java/com/pragma/plazoleta/infrastructure/exceptionhandler/ExceptionResponse.java
@@ -1,0 +1,15 @@
+package com.pragma.plazoleta.infrastructure.exceptionhandler;
+
+public enum ExceptionResponse {
+    NO_DATA_FOUND("No data found for the requested petition");
+
+    private final String message;
+
+    ExceptionResponse(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return this.message;
+    }
+}
