+++ src/main/java/com/pragma/plazoleta/infrastructure/input/rest/ObjectRestController.java
@@ -1,0 +1,53 @@
+package com.pragma.plazoleta.infrastructure.input.rest;
+
+import com.pragma.plazoleta.application.dto.request.ObjectRequestDto;
+import com.pragma.plazoleta.application.dto.response.ObjectResponseDto;
+import com.pragma.plazoleta.application.handler.IObjectHandler;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.ArraySchema;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/v1/object")
+@RequiredArgsConstructor
+public class ObjectRestController {
+
+    private final IObjectHandler objectHandler;
+
+    @Operation(summary = "Add a new object")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Object created", content = @Content),
+            @ApiResponse(responseCode = "409", description = "Object already exists", content = @Content)
+    })
+    @PostMapping("/")
+    public ResponseEntity<Void> saveObject(@RequestBody ObjectRequestDto objectRequestDto) {
+        objectHandler.saveObject(objectRequestDto);
+        return new ResponseEntity<>(HttpStatus.CREATED);
+    }
+
+    @Operation(summary = "Get all objects")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "All objects returned",
+                    content = @Content(mediaType = "application/json",
+                            array = @ArraySchema(schema = @Schema(implementation = ObjectResponseDto.class)))),
+            @ApiResponse(responseCode = "404", description = "No data found", content = @Content)
+    })
+    @GetMapping("/")
+    public ResponseEntity<List<ObjectResponseDto>> getAllObjects() {
+        return ResponseEntity.ok(objectHandler.getAllObjects());
+    }
+
+}
