--- src/main/java/com/pragma/powerup/PlazoletaApplication.java
@@ -1,13 +1,0 @@
-package com.pragma.powerup;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PlazoletaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PlazoletaApplication.class, args);
-	}
-
-}
