--- src/main/java/com/pragma/powerup/application/dto/request/ObjectRequestDto.java
@@ -1,10 +1,0 @@
-package com.pragma.powerup.application.dto.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ObjectRequestDto {
-    private String name;
-}
