--- src/main/java/com/pragma/powerup/application/dto/response/ObjectResponseDto.java
@@ -1,11 +1,0 @@
-package com.pragma.powerup.application.dto.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ObjectResponseDto {
-    private Long id;
-    private String name;
-}
