--- src/main/java/com/pragma/powerup/application/mapper/IObjectRequestMapper.java
@@ -1,13 +1,0 @@
-package com.pragma.powerup.application.mapper;
-
-import com.pragma.powerup.application.dto.request.ObjectRequestDto;
-import com.pragma.powerup.domain.model.ObjectModel;
-import org.mapstruct.Mapper;
-import org.mapstruct.ReportingPolicy;
-
-@Mapper(componentModel = "spring",
-        unmappedTargetPolicy = ReportingPolicy.IGNORE,
-        unmappedSourcePolicy = ReportingPolicy.IGNORE)
-public interface IObjectRequestMapper {
-    ObjectModel toObject(ObjectRequestDto objectRequestDto);
-}
