--- src/main/java/com/pragma/powerup/application/mapper/IObjectResponseMapper.java
@@ -1,17 +1,0 @@
-package com.pragma.powerup.application.mapper;
-
-import com.pragma.powerup.application.dto.response.ObjectResponseDto;
-import com.pragma.powerup.domain.model.ObjectModel;
-import org.mapstruct.Mapper;
-import org.mapstruct.ReportingPolicy;
-
-import java.util.List;
-
-@Mapper(componentModel = "spring",
-        unmappedTargetPolicy = ReportingPolicy.IGNORE,
-        unmappedSourcePolicy = ReportingPolicy.IGNORE)
-public interface IObjectResponseMapper {
-    ObjectResponseDto toResponse(ObjectModel objectModel);
-
-    List<ObjectResponseDto> toResponseList(List<ObjectModel> objectModelList);
-}
