--- src/main/java/com/pragma/powerup/domain/exception/DomainException.java
@@ -1,7 +1,0 @@
-package com.pragma.powerup.domain.exception;
-
-public class DomainException extends RuntimeException {
-    public DomainException(String message) {
-        super(message);
-    }
-}
