--- src/main/java/com/pragma/powerup/domain/model/ObjectModel.java
@@ -1,15 +1,0 @@
-package com.pragma.powerup.domain.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class ObjectModel {
-    private Long id;
-    private String name;
-}
