--- src/main/java/com/pragma/powerup/domain/spi/IObjectPersistencePort.java
@@ -1,10 +1,0 @@
-package com.pragma.powerup.domain.spi;
-
-import com.pragma.powerup.domain.model.ObjectModel;
-import java.util.List;
-
-public interface IObjectPersistencePort {
-    ObjectModel saveObject(ObjectModel objectModel);
-
-    List<ObjectModel> getAllObjects();
-}
