--- src/main/java/com/pragma/powerup/domain/usecase/ObjectUseCase.java
@@ -1,26 +1,0 @@
-package com.pragma.powerup.domain.usecase;
-
-import com.pragma.powerup.domain.api.IObjectServicePort;
-import com.pragma.powerup.domain.model.ObjectModel;
-import com.pragma.powerup.domain.spi.IObjectPersistencePort;
-
-import java.util.List;
-
-public class ObjectUseCase implements IObjectServicePort {
-
-    private final IObjectPersistencePort objectPersistencePort;
-
-    public ObjectUseCase(IObjectPersistencePort objectPersistencePort) {
-        this.objectPersistencePort = objectPersistencePort;
-    }
-
-    @Override
-    public void saveObject(ObjectModel objectModel) {
-        objectPersistencePort.saveObject(objectModel);
-    }
-
-    @Override
-    public List<ObjectModel> getAllObjects() {
-        return objectPersistencePort.getAllObjects();
-    }
-}
