--- src/main/java/com/pragma/powerup/infrastructure/exception/NoDataFoundException.java
@@ -1,7 +1,0 @@
-package com.pragma.powerup.infrastructure.exception;
-
-public class NoDataFoundException extends RuntimeException {
-    public NoDataFoundException() {
-        super();
-    }
-}
