--- src/main/java/com/pragma/powerup/infrastructure/exceptionhandler/ExceptionResponse.java
@@ -1,15 +1,0 @@
-package com.pragma.powerup.infrastructure.exceptionhandler;
-
-public enum ExceptionResponse {
-    NO_DATA_FOUND("No data found for the requested petition");
-
-    private final String message;
-
-    ExceptionResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return this.message;
-    }
-}
