+++ src/test/java/com/pragma/plazoleta/PowerUpApplicationTests.java
@@ -1,0 +1,13 @@
+package com.pragma.plazoleta;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PowerUpApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
