--- src/test/java/com/pragma/powerup/PowerUpApplicationTests.java
@@ -1,13 +1,0 @@
-package com.pragma.powerup;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PowerUpApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
